requirements.txt
fastapi
uvicorn
streamlit
langchain
groq
textblob
python-dotenv
plotly